feat: make stream types more extensible
