+++ arcon/src/metrics/perf_event.rs
@@ -74,8 +74,7 @@ impl fmt::Display for HardwareCounter {
 ///                 ..Default::default()
 ///             },
 ///         })
-///         .to_console()
-///         .build();
+///         .to_console();
 ///     app.start();
 ///     app.await_termination();
 /// }
