+++ examples/file.rs
@@ -5,9 +5,8 @@ fn main() {
         .file("file_source_data", |cfg| {
             cfg.set_arcon_time(ArconTime::Process);
         })
-        .flatmap(|x| (0..x))
-        .to_console()
-        .build();
+        .flat_map(|x| (0..x))
+        .to_console();
 
     app.start();
     app.await_termination();
